+++ jina/helper.py
@@ -25,7 +25,8 @@ __all__ = ['batch_iterator', 'yaml',
            'parse_arg',
            'random_port', 'get_random_identity', 'expand_env_var',
            'colored', 'kwargs2list', 'get_local_config_source', 'is_valid_local_config_source',
-           'cached_property', 'is_url', 'complete_path', 'typename', 'get_public_ip', 'get_internal_ip']
+           'cached_property', 'is_url', 'complete_path',
+           'typename', 'get_public_ip', 'get_internal_ip', 'guess_mime']
 
 
 def deprecated_alias(**aliases):
